Refactoring to one JS package
