--- packages/javascript/jupyterlab-plotly/README.md
@@ -1,5 +1,0 @@
-JupyterLab Extension for Plotly.py
-
-See https://plot.ly/python/getting-started/ for details.
-
-Install with `jupyter labextension install jupyterlab-plotly`
