--- packages/javascript/jupyterlab-plotly/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "jupyterlab-plotly",
-  "version": "4.14.3",
-  "description": "The plotly JupyterLab extension",
-  "author": "The plotly.py team",
-  "license": "MIT",
-  "main": "src/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/plotly/plotly.py"
-  },
-  "keywords": [
-    "jupyter",
-    "ipython",
-    "plotly"
-  ],
-  "files": [
-    "src/**/*.js",
-    "dist/*.js",
-    "style/*.*"
-  ],
-  "scripts": {
-    "build": "npm run build:src",
-    "build:src": "rimraf dist && tsc",
-    "clean": "rimraf dist/",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "devDependencies": {
-    "rimraf": "^2.6.1",
-    "ify-loader": "^1.1.0",
-    "typescript": "~3.7.0"
-  },
-  "dependencies": {
-    "plotly.js": "^1.58.4",
-    "@types/plotly.js": "1.44.28",
-    "@jupyterlab/rendermime-interfaces": "^1.3.0 || ^2.0.0 || ^3.0.0",
-    "@lumino/messaging": "^1.2.3",
-    "@lumino/widgets": "^1.8.1",
-    "lodash": "^4.17.4"
-  },
-  "jupyterlab": {
-    "mimeExtension": "dist/javascript-renderer-extension.js"
-  }
-}
