--- packages/javascript/jupyterlab-plotly/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "declaration": true,
-    "noImplicitAny": true,
-    "noEmitOnError": true,
-    "noUnusedLocals": true,
-    "esModuleInterop": true,
-    "preserveWatchOutput": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "target": "es2015",
-    "lib": ["dom", "es2015"],
-    "types": []
-  },
-  "include": ["src/*"]
-}
