+++ packages/javascript/plotlywidget/package.json
@@ -1,10 +1,10 @@
 {
   "name": "plotlywidget",
   "version": "4.14.3",
-  "description": "The plotly JupyterLab extension",
+  "description": "The plotly Jupyter extension",
   "author": "The plotly.py team",
   "license": "MIT",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/plotly/plotly.py"
@@ -17,28 +17,60 @@
     "plotly"
   ],
   "files": [
-    "src/**/*.js",
+    "lib/**/*.js",
     "dist/*.js",
     "style/*.*"
   ],
   "scripts": {
-    "build": "webpack",
-    "clean": "rimraf dist/ && rimraf  ../../python/plotly/plotlywidget/static'",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "npm run build:lib && npm run build:nbextension && npm run build:labextension:dev",
+    "build:prod": "npm run build:lib && npm run build:nbextension && npm run build:labextension",
+    "build:labextension": "jupyter labextension build .",
+    "build:labextension:dev": "jupyter labextension build --development True .",
+    "build:lib": "tsc",
+    "build:nbextension": "webpack --mode=production",
+    "clean": "npm run clean:lib && npm run clean:nbextension && npm run clean:labextension",
+    "clean:lib": "rimraf lib",
+    "clean:labextension": "rimraf ../../python/plotly/plotlywidget/labextension",
+    "clean:nbextension": "rimraf ../../python/plotly/plotlywidget/nbextension/index.js*",
+    "lint": "eslint . --ext .ts,.tsx --fix",
+    "lint:check": "eslint . --ext .ts,.tsx",
+    "prepack": "npm run build:lib",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "npm-run-all -p watch:*",
+    "watch:lib": "tsc -w",
+    "watch:nbextension": "webpack --watch"
   },
   "devDependencies": {
-    "webpack": "^3.10.0",
-    "rimraf": "^2.6.1",
-    "ify-loader": "^1.1.0",
-    "typescript": "~3.1.1"
+    "@jupyterlab/builder": "^3.0.0",
+    "@lumino/application": "^1.6.0",
+    "@types/plotly.js": "^1.54.10",
+    "@types/webpack-env": "^1.13.6",
+    "acorn": "^7.2.0",
+    "css-loader": "^3.2.0",
+    "fs-extra": "^7.0.0",
+    "mkdirp": "^0.5.1",
+    "npm-run-all": "^4.1.3",
+    "prettier": "^2.0.5",
+    "rimraf": "^2.6.2",
+    "source-map-loader": "^1.1.3",
+    "style-loader": "^1.0.0",
+    "ts-loader": "^8.0.0",
+    "typescript": "~4.1.3",
+    "webpack": "^5.0.0",
+    "webpack-cli": "^4.0.0"
   },
   "dependencies": {
-    "plotly.js": "^1.58.4",
     "@jupyter-widgets/base": "^2.0.0 || ^3.0.0 || ^4.0.0",
-    "lodash": "^4.17.4"
+    "@jupyterlab/rendermime-interfaces": "^1.3.0 || ^2.0.0 || ^3.0.0",
+    "@lumino/messaging": "^1.2.3",
+    "@lumino/widgets": "^1.8.1",
+    "lodash": "^4.17.4",
+    "plotly.js": "^1.58.4"
   },
   "jupyterlab": {
-    "extension": "src/jupyterlab-plugin.js",
+    "extension": "lib/jupyterlab-plugin",
+    "mimeExtension": "lib/plotly-renderer",
+    "outputDir": "../../python/plotly/plotlywidget/labextension",
     "sharedPackages": {
       "@jupyter-widgets/base": {
         "bundled": false,
