+++ packages/javascript/plotlywidget/src/extension.ts
@@ -1,0 +1,13 @@
+// Entry point for the notebook bundle containing custom model definitions.
+//
+// Setup notebook base URL
+//
+// Some static assets may be required by the custom widget javascript. The base
+// url for the notebook is not known at build time and is therefore computed
+// dynamically.
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+(window as any).__webpack_public_path__ =
+  document.querySelector('body')!.getAttribute('data-base-url') +
+  'nbextensions/plotlywidget';
+
+export * from './index';
