--- packages/javascript/plotlywidget/src/index.js
@@ -1,3 +1,0 @@
-// Export widget models and views, and the npm package version number.
-module.exports = require('./Figure.js');
-module.exports['version'] = require('../package.json').version;
