+++ packages/javascript/plotlywidget/src/index.ts
@@ -1,0 +1,3 @@
+// Export widget models and views, and the npm package version number.
+export * from './Figure';
+export * from './version';
