--- packages/javascript/plotlywidget/src/jupyterlab-plugin.js
@@ -1,18 +1,0 @@
-var plotly = require('./index');
-var base = require('@jupyter-widgets/base');
-
-/**
- * The widget manager provider.
- */
-module.exports = {
-  id: 'plotlywidget',
-  requires: [base.IJupyterWidgetRegistry],
-  activate: function(app, widgets) {
-      widgets.registerWidget({
-          name: 'plotlywidget',
-          version: plotly.version,
-          exports: plotly
-      });
-    },
-  autoStart: true
-};
