+++ packages/python/plotly/plotlywidget/nbextension/extension.js
@@ -1,0 +1,18 @@
+
+// Entry point for the notebook bundle containing custom model definitions.
+//
+define(function() {
+    "use strict";
+
+    window['requirejs'].config({
+        map: {
+            '*': {
+                'plotlywidget': 'nbextensions/plotlywidget/index',
+            },
+        }
+    });
+    // Export the required load_ipython_extension function
+    return {
+        load_ipython_extension : function() {}
+    };
+});
